requirements.txt
python-constraint2==2.0.0b4
simanneal==0.5.0